chore: extract shared code from sidebar
